Crate ecvrf

Implementation of the verifiable random function as specified in https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-vrf-09.

Structs

Keypair

An ed25519 keypair.

Proof

Proof that the value produced by the VRF function is correct with respect to a given public key.

ProofError

Errors which may occur while processing proofs and keypairs.

PublicKey

An ed25519-like public key. This has a bit stricter requirements than the signature scheme public keys, in particular points of small order are not allowed, and this is checked during serialization.

SecretKey

An EdDSA secret key.

Constants

EXPANDED_SECRET_KEY_LENGTH

The length of an “expanded” ed25519 key, ExpandedSecretKey, in bytes.

KEYPAIR_LENGTH

The length of an ed25519 Keypair, in bytes.

ONE_STRING

PROOF_LENGTH

The length of a VRF Proof, in bytes.

PUBLIC_KEY_LENGTH

The length of an ed25519 PublicKey, in bytes.

SECRET_KEY_LENGTH

The length of a ed25519 SecretKey, in bytes.

SUITE_STRING

Suite string as defined by https://tools.ietf.org/id/draft-irtf-cfrg-vrf-07.html#suites

THREE_STRING

TWO_STRING

ZERO_STRING

Strings combined with inputs to the hash function for domain separation, as discussed in https://tools.ietf.org/id/draft-irtf-cfrg-vrf-07.html#rfc.section.7.7

Functions

hash_points

Implements https://tools.ietf.org/id/draft-irtf-cfrg-vrf-07.html#rfc.section.5.4.3